# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Movie API
  version: 1.0.0
  description: API de gestion des films + genres

servers:
  - url: http://localhost:8082

tags:
  - name: movies
    description: Gestion des films
  - name: genres
    description: Gestion des genres

paths:
  /api/movies:
    get:
      tags: [movies]
      summary: Liste des films
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MovieDto"
    post:
      tags: [movies]
      summary: Créer un film
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieDto"
      responses:
        "200":
          description: OK (créé)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDto"

  /api/movies/{id}:
    get:
      tags: [movies]
      summary: Détail film
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDto"
        "404":
          description: Non trouvé
    put:
      tags: [movies]
      summary: Modifier film
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieDto"
      responses:
        "200":
          description: OK (modifié)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDto"

  /api/movies/{id}/open:
    put:
      tags: [movies]
      summary: Ouvrir un film à la location (avec prix)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: prix
          in: query
          required: true
          schema: { type: number, format: double }
      responses:
        "200":
          description: Film ouvert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDto"

  /api/movies/{id}/close:
    put:
      tags: [movies]
      summary: Fermer un film à la location
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Film fermé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDto"

  /api/genres:
    get:
      tags: [genres]
      summary: Liste des genres
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Genre"
    post:
      tags: [genres]
      summary: Créer un genre (optionnel/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenreCreateRequest"
      responses:
        "200":
          description: OK (créé)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Genre"

  /api/genres/{id}:
    get:
      tags: [genres]
      summary: Détail genre
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Genre"
        "404":
          description: Non trouvé
    delete:
      tags: [genres]
      summary: Supprimer un genre (optionnel)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean

components:
  schemas:
    MovieDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        titre:
          type: string
        annee:
          type: integer
          nullable: true
        ageMinimum:
          type: integer
          nullable: true
        ouvertLocation:
          type: boolean
          nullable: true
        prixLocation:
          type: number
          format: double
          nullable: true
        realisateurId:
          type: integer
          format: int64
          nullable: true
        acteurIds:
          type: array
          items:
            type: integer
            format: int64
          nullable: true
        genres:
          type: array
          items:
            type: string
          nullable: true
      required:
        - titre

    Genre:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      required:
        - name

    GenreCreateRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name